+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Created a HoverableNavigationRailDestination Widget that extends the NavigationRailDestination Widget which accepts a Function(bool isHovered) onHoverStateChange
+* Created a HoverableNavigationRail Widget which can handle the HoverableNavigationRailDestination's onHoverStateChange